Time: 0 ms (100.00%), Space: 41.8 MB (6.09%) - LeetHub

+++ 62-unique-paths/62-unique-paths.java
@@ -2,9 +2,9 @@ class Solution {
     long store[];
     public int uniquePaths(int m, int n) {
         int k = Math.max(n,m);
-        long val =1;
-        for(int i=k,c=1;i<=(m+n-2);i++,c++) {
-            val=val * i/c;
+        double val =1;
+        for(double i=k,c=1;i<=(m+n-2);i++,c++) {
+            val=(val * i)/c;
             //val/=c;
         }
         return (int)val;
